+++ helion/language/creation_ops.py
@@ -149,25 +149,6 @@ def _full_codegen(state: CodegenState) -> ast.AST:
     )
 
 
-@_decorators.codegen(full, "pallas")
-def _full_codegen_pallas(state: CodegenState) -> ast.AST:
-    """Pallas codegen for hl.full / hl.zeros.
-
-    Always lowers to a plain ``jnp.full`` bound to a fresh local, regardless
-    of pallas_loop_type.  The previous emit_pipeline/fori_loop path returned
-    a bare scratch ref AST, which broke any downstream arithmetic on the
-    result outside the inner loop -- e.g. ``acc = hl.zeros(...); acc += x``
-    emitted ``scratch + x`` and JAX raised
-    ``'AbstractRef' object has no attribute '_add'`` at trace time
-    (``Refs`` don't support arithmetic; only ``ref[...]`` reads).
-
-    When the result is loop-carried, ``_setup_loop_carried_state`` allocates
-    a scratch buffer at the loop boundary and copies the init value in --
-    no scratch needed at the ``hl.zeros`` site itself.
-    """
-    return full._codegen["common"](state)  # pyrefly: ignore[missing-attribute]
-
-
 @_decorators.get_masked_value(full)
 def _(
     node: torch.fx.Node,
@@ -237,3 +218,11 @@ def arange(
         dtype=dtype,
         device=env.device if device is None else device,
     )
+
+
+# ---------------------------------------------------------------------------
+# Backend-specific codegens for these ops live in per-backend modules under
+# helion/_compiler/<backend>/.  Import them here (at module import time) so the
+# @_decorators.codegen(op, "<backend>") registrations run with the same eager
+# timing as when the bodies lived in this file -- no behavior change.
+import helion._compiler.pallas.creation_ops  # noqa: E402, F401
